[Matrix] Fix constant index swizzle
fixes #172805

We need to be calling `MakeExtVectorElt` in `EmitExtVectorElementExpr` so
that we can call `E->getEncodedElementAccess(Elts);` off of the `ExtVectorElementExpr`
Node.

Calling it in `EmitMatrixSingleSubscriptExpr` was wrong because we doing
have the vector encoded elementts yet. But also to get the right index
we need a way to pass on the matrix column and row lengths so
`MakeMatrixSubVector` needed to be expanded to store row and column
sizes on the LValue.

It also is looking like we will need to revisit the indexing scheme
after #171564 because the single index seems to also be highly dependend
on row\col major order.